+++ Project-Elara/elara-admin-console/src/main/java/ru/hse/elarateam/admin_console/ElaraAdminConsoleApplication.java
@@ -1,0 +1,11 @@
+package ru.hse.elarateam.admin_console;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ElaraAdminConsoleApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ElaraAdminConsoleApplication.class, args);
+    }
+}
