+++ Project-Elara/elara-admin-console/src/test/java/ru/hse/elarateam/admin_console/ElaraAdminConsoleApplicationTest.java
@@ -1,0 +1,13 @@
+package ru.hse.elarateam.admin_console;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ElaraAdminConsoleApplicationTest {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
