+++ Project-Elara/elara-products/src/main/java/ru/hse/elarateam/products/model/Picture.java
@@ -1,4 +1,4 @@
-package ru.hse.products.model;
+package ru.hse.elarateam.products.model;
 
 import jakarta.persistence.*;
 import lombok.*;
