+++ Project-Elara/elara-products/src/main/java/ru/hse/elarateam/products/repositories/ProductsRepository.java
@@ -1,11 +1,11 @@
-package ru.hse.products.repositories;
+package ru.hse.elarateam.products.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-import ru.hse.products.model.Product;
+import ru.hse.elarateam.products.model.Product;
 
 import java.util.UUID;
