+++ Project-Elara/elara-products/src/main/java/ru/hse/elarateam/products/web/controllers/ProductsControllerExceptionHandler.java
@@ -1,4 +1,4 @@
-package ru.hse.products.web.controllers;
+package ru.hse.elarateam.products.web.controllers;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
